pyproject.toml
[tool.poetry]
name = "directus_requests"
version = "0.0.1"
description = "A repository to group all little scripts that are not made to be automated afterwards. For example, scripts to update only one time data in directus, but not scripts that will automate a process."
authors = ["Edouard Brülhart <fedouard.bruelhart@unifr.ch>"]
repository = "https://github.com/edouardbruelhart/directus-requests"
documentation = "https://edouardbruelhart.github.io/directus-requests/"
readme = "README.md"
packages = [
  {include = "directus_requests"}
]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
requests = "^2.31.0"
types-requests = "^2.31.0.20240406"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
mypy = "^1.5.1"
pre-commit = "^3.4.0"
tox = "^4.11.1"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = ["directus_requests"]
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py37"
line-length = 120
fix = true
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.format]
preview = true



[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]
